Lottery egld optimizations

+++ contracts/examples/lottery-egld/src/lottery_info.rs
@@ -11,6 +11,6 @@ pub struct LotteryInfo<BigUint: BigUintApi> {
 	pub max_entries_per_user: u32,
 	pub prize_distribution: Vec<u8>,
 	pub whitelist: Vec<Address>,
-	pub current_ticket_number: u32,
+	pub current_ticket_number: usize,
 	pub prize_pool: BigUint,
 }
